added concurrency, error codes, versioning, and miscellaneous files, removed corresponding tests
